Extract HomePageComponent
